PM-31 refactored pet list and pet filters
