+++ src/components/Header.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
-import AuthContext from '../../context/AuthContext';
+import AuthContext from '../context/AuthContext';
 
 export default class Header extends Component {
   static contextType = AuthContext;
