+++ src/routes/PetsPage/PetsPage.js
@@ -1,14 +1,13 @@
 import React, { Component } from 'react';
-import Select from 'react-select';
-import PetCard from '../../components/PetCard/PetCard';
-import { Input, Section } from '../../components/Utils/Utils';
+import PetFilters from '../../components/PetFilters';
+import PetList from '../../components/PetList';
+import { Section } from '../../components/Utils/Utils';
 import PetsService from '../../services/petsService';
 
 export default class PetsPage extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      breeds: [],
       pets: [],
       petCount: -1, // this is set to prevent the no pets alert before data is loaded
       typeOfAnimal: [],
@@ -18,12 +17,11 @@ export default class PetsPage extends Component {
       maxAge: '',
       more: []
     };
+    this.inputChanged = this.inputChanged.bind(this);
+    this.multiSelectChanged = this.multiSelectChanged.bind(this);
   }
 
   componentDidMount() {
-    PetsService.getBreeds()
-      .then(breeds => this.setState({ breeds }))
-      .catch(error => console.log(error));
     PetsService.getPets({})
       .then(pets => this.setState({ pets, petCount: pets.length }))
       .catch(error => console.log(error));
@@ -34,7 +32,6 @@ export default class PetsPage extends Component {
   }
 
   multiSelectChanged(field, selectedOptions) {
-    const values = Array.from(selectedOptions, option => option.value);
     this.setState({ [field]: selectedOptions }, () => this.handleSubmit());
   }
 
@@ -71,10 +68,11 @@ export default class PetsPage extends Component {
     more.forEach(option => filters[option.value] = true);
     return filters;
   }
-
-  renderFilters() {
+  
+  render() {
     const {
-      breeds,
+      pets,
+      petCount,
       typeOfAnimal,
       breed,
       sex,
@@ -82,141 +80,20 @@ export default class PetsPage extends Component {
       maxAge,
       more
     } = this.state;
-    const typeOptions = [
-      { value: 'Cat', label: 'Cat' },
-      { value: 'Dog', label: 'Dog' },
-      { value: 'Other', label: 'Other' }
-    ];
-    const breedOptions = breeds
-      .filter(breed => typeOfAnimal.length > 0
-        ? typeOfAnimal.find(type => type.value === breed.TypeOfAnimal) : true
-      )
-      .map(breed => {
-        return { value: breed.Breed, label: `${breed.TypeOfAnimal} - ${breed.Breed}` };
-      });
-    const sexOptions = [
-      { value: 'Female', label: 'Female' },
-      { value: 'Male', label: 'Male' }
-    ];
-    const moreOptions = [
-      { value: 'goodWithOtherAnimals', label: 'Good With Other Animals' },
-      { value: 'goodWithChildren', label: 'Good With Children' },
-      { value: 'leashNotRequired', label: 'Leash Not Required' },
-      { value: 'neutered', label: 'Neutered' },
-      { value: 'vaccinated', label: 'Vaccinated' },
-      { value: 'houseTrained', label: 'House Trained' },
-    ];
-    return (
-      <form>
-        <div className='row mb-2'>
-          <div className='col-sm m-1'>
-            <Select
-              placeholder='TYPE OF PET'
-              name='typeOfAnimal'
-              id='typeOfAnimal'
-              value={typeOfAnimal}
-              options={typeOptions}
-              onChange={selectedOptions => this.multiSelectChanged('typeOfAnimal', selectedOptions)}
-              isMulti
-            />
-          </div>
-          <div className='col-sm m-1'>
-            <Select
-              placeholder='BREED'
-              name='breed'
-              id='breed'
-              value={breed}
-              options={breedOptions}
-              onChange={selectedOptions => this.multiSelectChanged('breed', selectedOptions)}
-              isMulti
-            />
-          </div>
-          <div className='col-sm m-1'>
-            <Select
-              placeholder='SEX'
-              name='sex'
-              id='sex'
-              value={sex}
-              options={sexOptions}
-              onChange={selectedOption => this.inputChanged('sex', selectedOption)}
-              isClearable
-            />
-          </div>
-        </div>
-        <div className='row mb-2'>
-          <div className='col-sm m-1'>
-            <Input
-              className=''
-              placeholder='FROM AGE'
-              name='minAge'
-              id='minAge'
-              type='number'
-              value={minAge}
-              onChange={e => this.inputChanged('minAge', e.target.value)}
-            />
-          </div>
-          <div className='col-sm m-1'>
-            <Input
-              className=''
-              placeholder='TO AGE'
-              name='maxAge'
-              id='maxAge'
-              type='number'
-              value={maxAge}
-              onChange={e => this.inputChanged('maxAge', e.target.value)}
-            />
-          </div>
-          <div className='col-sm m-1'>
-            <Select
-              placeholder='MORE'
-              name='more'
-              id='more'
-              value={more}
-              options={moreOptions}
-              onChange={selectedOptions => this.multiSelectChanged('more', selectedOptions)}
-              isMulti
-            />
-          </div>
-        </div>
-      </form>
-    );
-  }
-
-  renderPetRow(rowStart) {
-    return (
-      <div key={rowStart} className='row mb-2'>
-        {this.state.pets.slice(rowStart, rowStart + 3).map(pet => this.renderPet(pet))}
-      </div>
-    );
-  }
-
-  renderPet(pet) {
-    return (
-      <div key={pet.PetID} className='col-sm-4'>
-        <PetCard pet={pet} page='pets' className='border rounded m-1 p-2' />
-      </div>
-    );
-  }
-  
-  render() {
-    const { petCount } = this.state;
-
-    // no pets exist
-    if (petCount === 0) {
-      return (
-        <div className='alert alert-success' role='alert'>
-          More pets will be available soon! Check back later :)
-        </div>
-      );
-    }
-
-    // place pets in a 3-cols grid
-    const petRows = Math.floor(petCount / 3) + (petCount % 3 === 0 ? 0 : 1);
     return (
       <Section>
-        {this.renderFilters()}
+        <PetFilters
+          typeOfAnimal={typeOfAnimal}
+          breed={breed}
+          sex={sex}
+          minAge={minAge}
+          maxAge={maxAge}
+          more={more}
+          inputChangeHandler={this.inputChanged}
+          multiSelectChangeHandler={this.multiSelectChanged}
+        />
         <br />
-        {[...Array(petRows)].map((_, i) => this.renderPetRow(i * 3))}
+        <PetList pets={pets} petCount = {petCount} page='pets' />
       </Section>
     );
   }
